requirements.txt
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0

torch>=2.0.0
torchvision>=0.15.0

transformers>=4.35.0
accelerate>=0.25.0

jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=7.0.0

python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0

matplotlib>=3.7.0
seaborn>=0.13.0

sentence-transformers>=2.2.0
faiss-cpu>=1.7.4